Update and rename untitled.py to app.py
Update
